Rework the module to utilize inheritable Proxy
Thanks to @lizmat work!

This implies Rakudo of at least v2019.11, unfortunately. But this move
pays of as it allows to get around a bug in MoarVM which causes
`DESTROY`s not always be called.

Also, it provides better support for concurrency without additional
locks required. Of course, it doesn't mean that `$inst.attr = $val`
would behave better than without `mooish` trait. It only means that it
won't behave worse!

Removed tests for `mooish-obj-count` because this routine doesn't make
sense anymore.

Fixed a few test glitches. Mostly related to optimized away calls to
`FETCH` on `Proxy`.